rename component names
