+++ src/components/MediaPlayer/AudioInfo/AudioInfo.scss
