+++ src/components/MediaPlayer/MainControl/Buttons/Buttons.scss
