+++ src/components/MediaPlayer/MainControl/MainControl.scss
@@ -1,5 +1,5 @@
-@import "./Buttons/ButtonsComponent";
-@import "./VolumeBar/VolumeBarComponent";
+@import "./Buttons/Buttons";
+@import "./VolumeBar/VolumeBar";
 
 .main-control {
   display: flex;
