+++ src/components/MediaPlayer/MainControl/VolumeBar/VolumeBar.scss
