+++ src/components/MediaPlayer/ProgressBar/ProgressBar.scss
