+++ src/components/MediaPlayer/Rating/Rating.scss
