+++ src/index.scss
@@ -12,7 +12,7 @@
 @import "./components/Header/Header";
 @import "./components/Search/Search";
 @import "./components/Main/Main";
-@import "./components/MediaPlayer/MediaPlayerComponent";
+@import "./components/MediaPlayer/MediaPlayer";
 @import "./components/Page/Page";
 @import "./components/About/About";
 @import "./components/DotsMenu/DotsMenu";
